requirements.txt
# general modules
Django>=4.2,<4.3
python-decouple==3.8
pillow==10.2.0
djangorestframework

# third party modules 
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt

# deployment modules


# email third party modules 
django-mail-templated

# reformater and test modules
black
flake8
